+++ src/app/(route)/(community)/news/page.tsx
@@ -4,10 +4,10 @@ import React from "react";
 import NewsListContent from "./_components/NewsListContent";
 import useSortedNewsDataList from "@/_hooks/useNews/useSortedPosts";
 import NewsTalkToolbar from "./_components/NewsTalkToolbar";
-import { useNewsListLogic } from "@/_hooks/useNews/useNewsListLogic";
+import { useNewsPageLogic } from "@/utils/newsUtils/useNewsPageLogic";
 
 const Page = () => {
-  const { orderType, setOrderType, pageNum, onPageChange } = useNewsListLogic();
+  const { orderType, setOrderType, pageNum, onPageChange } = useNewsPageLogic();
   const {
     data: newsListData,
     isLoading,
