+++ src/app/(route)/main/_components/MainBigSizeNews.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import Image from "next/image";
 import useGetNewsDataList from "@/_hooks/useNews/useGetNewsDataList";
 import { useRouter } from "next/navigation";
-import { useReadNews } from "@/_hooks/useReadNews";
+import { useReadNews } from "@/_hooks/useNews/useReadNews";
 
 const MainBigSizeNews = () => {
   const router = useRouter();
