+++ src/utils/newsUtils/changedCategory.ts
@@ -1,5 +1,3 @@
-import React from "react";
-
 interface ChangedCategoryProps {
   category?: string;
 }
