+++ src/utils/newsUtils/useNewsPageLogic.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-export const useNewsListLogic = () => {
+export const useNewsPageLogic = () => {
   const [orderType, setOrderType] = useState<"DATE" | "COMMENT" | "VIEW">("DATE");
   const [pageNum, setPageNum] = useState(1);
